einiges
LH Anpassungen

WayProvider im Grunde fertig, Lufthansa macht Faxen

Zug- und Flugzeugwege werden zurückgegeben

WaysProvider angepasst

sehr unfertig

+++ src/main/java/travelcompare/restapi/external/viamichelin/model/RoadSheetStep.java
@@ -11,7 +11,7 @@ public class RoadSheetStep {
 
     @Getter
     @Setter
-    private Long duration;
+    private double duration;
 
     @Getter
     @Setter
